sources/requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Assets==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
neo4j==5.24.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
waitress==3.0.0
webassets==2.0
Werkzeug==3.0.4
